Całe mnóstwo zmian: poprawiono tworzenie portali (nie zmienia ID o ile to możliwe); zabezpieczenie przy usuwaniu elementów; niemożliwa modyfikacja sieci gdy trwa symulacja; poprawiono backup główny m0; wprowadzono backup wewnętrzny StateSimulator (czyli nie namiesza ani nie zresetuje to, co aktualnie ustawił symulator główny, jeśli cokolwiek). Inne drobne.

+++ DarkAbyss/src/abyss/graphpanel/popupmenu/PetriNetElementPopupMenu.java
@@ -3,6 +3,8 @@ package abyss.graphpanel.popupmenu;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+import javax.swing.JOptionPane;
+
 import abyss.darkgui.GUIManager;
 import abyss.graphpanel.GraphPanel;
 import abyss.math.Node;
@@ -28,18 +30,18 @@ public class PetriNetElementPopupMenu extends GraphPanelPopupMenu {
 		
 		this.addMenuItem("Show details...", "", new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				//NetSimulator ns = GUIManager.getDefaultGUIManager().getSimulatorBox().getCurrentDockWindow().getSimulator();
-				//if(ns.getMode() == SimulatorMode.STOPPED) {
+				if(getGraphPanel().getSelectionManager().getSelectedElementLocations().size() == 0)
+					return;
 				
-				//TODO: wykrywanie kliknięcia łuku, inaczej wylatuje wyjątek:
-					Node n = getGraphPanel().getSelectionManager().getSelectedElementLocations().get(0).getParentNode();
-					if(n instanceof Place) {
-						AbyssNodeInfo ani = new AbyssNodeInfo((Place)n, GUIManager.getDefaultGUIManager().getFrame());
-						ani.setVisible(true);
-					} else if(n instanceof Transition) {
-						AbyssNodeInfo ani = new AbyssNodeInfo((Transition)n, GUIManager.getDefaultGUIManager().getFrame());
-						ani.setVisible(true);
-					} 
+				Node n = getGraphPanel().getSelectionManager().getSelectedElementLocations().get(0).getParentNode();
+				if(n instanceof Place) {
+					AbyssNodeInfo ani = new AbyssNodeInfo((Place)n, GUIManager.getDefaultGUIManager().getFrame());
+					ani.setVisible(true);
+				} else if(n instanceof Transition) {
+					AbyssNodeInfo ani = new AbyssNodeInfo((Transition)n, GUIManager.getDefaultGUIManager().getFrame());
+					ani.setVisible(true);
+				} 
+				
 				//} else {
 				//	JOptionPane.showMessageDialog(null, "Warning: simulator active. Cannot proceed until manually stopped.",
 				//			"Net simulator working", JOptionPane.WARNING_MESSAGE);
@@ -49,8 +51,22 @@ public class PetriNetElementPopupMenu extends GraphPanelPopupMenu {
 		
 		this.addMenuItem("Delete", "cross.png", new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				//TODO: warning question
-				getGraphPanel().getSelectionManager().deleteAllSelectedElements();
+				if(GUIManager.getDefaultGUIManager().reset.isSimulatorActiveWarning(
+						"Operation impossible when simulator is working."
+						, "Warning") == true)
+					return;
+				
+				Object[] options = {"Delete", "Cancel",};
+				int n = JOptionPane.showOptionDialog(null,
+						"Do you want to delete selected elements?",
+						"Deletion warning?", JOptionPane.YES_NO_OPTION,
+						JOptionPane.QUESTION_MESSAGE, null, options, options[0]);
+				if (n == 0) {
+					if(GUIManager.getDefaultGUIManager().getWorkspace().getProject().isBackup == true) {
+						GUIManager.getDefaultGUIManager().getWorkspace().getProject().restoreMarkingZero();
+					}
+					getGraphPanel().getSelectionManager().deleteAllSelectedElements();
+				}
 			}
 		});
